rename pipeline to workflow. move expandJson class to expandWorkflow
