+++ hermes/__init__.py
@@ -24,8 +24,8 @@ __version__=(1,0,0)
 """
 
 
-from .pipeline.workflow import hermesWorkflow
+from .workflow.workflow import hermesWorkflow
 from .taskwrapper import hermesTaskWrapper
 from .Resources.nodeTemplates.templateCenter import templateCenter
-from .pipeline.expandWorkflow import expandWorkflow
+from .workflow.expandWorkflow import expandWorkflow
 from .Resources.executers.executerHome import executerHome
